Create a function to get user's choice in script.js
The function includes a while loop to get the user
choice because users may try to put invalid values
through or misspell something, in which case the
function should keep asking for a valid value (that
is: "rock", "paper", or "scissors") before returning
anything. The code is set up in a way that ensures
that all invalid values are either disregarded or
are checked, but do not break the while loop until
some variation of the valid values is inputted (for
example: "SciSSors" instead of "scissors"), which
would then lead the function to return that value.