Add logs to show score after round in script.js
At first, I tried to use a 'default' block in the
switch statement to log these messages, forgetting
that the 'default' block is effectively the 'else'
block in a traditional conditional statement. The
desired result was to show the score at the end of
a round regardless of the round's results, not show
them only when no winner could be decided (that makes
sense because all the cases before the 'default' block
handle all the possibilities that a user can win or
lose a round).