Fix condition to show final winner message in script.js
At first, regardless of whether the user wanted to
start the game or not (by typing 'yes' or 'no' in
the prompt box), the function would always show the
final winner message. Rewriting the code into a
conditional statement using the value inputted by
the user into the prompt as the condition made it
easy to put the final winner message in the code block
that would run ONLY if the user wanted to start a
game and show some other appropriate message otherwise.