Rewrite func. 'getGameEndStatus' to use played rounds
The function 'getGameEndStatus' was returning true
every time the total of the players' scores was one
less than the number of rounds, in anticipation of
the game ending. But just because the scores may
equal one less than the number of rounds, it doesn't
mean that that sum was the number of rounds played
up until that point. Using a round counter variable
such as 'numOfPlayedRounds' to count the number of
rounds each time a round is played out fixes this.
* This program is written such that only a select
number of rounds are played each game. *